requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
redis>=5.0.0
python-jose
pydantic
typing-extensions
stripe
asyncpg
psycopg2-binary
google-generativeai==0.1.0rc1

